net

Module async_runtime

Structs

AsyncRuntimeHolder

The actual runtime, to be used as part of shut-down.

Statics

ASYNC_RUNTIME_HANDLE 🔒

A shared handle to the runtime, to be initialized on start-up.

Functions

async_runtime_initialized

init_async_runtime

spawn_blocking_task

Spawn a blocking task using the handle to the runtime.

spawn_task

Spawn a task using the handle to the runtime.