net

Module connector

Structs

CertificateErrorOverrideManager

This data structure is used to track certificate verification errors and overrides. It tracks:

CertificateErrorOverrideManagerInternal 🔒

CertificateVerificationOverrideVerifier 🔒

InstrumentedConnector

InstrumentedStream

ProxyConnector

A proxy connector. This will automatically open a proxy connection if the uri matches the proxy uri. Also respects ‘no_proxy’.

ServoHttpConnector

TlsHandshakeInfo

TokioExecutor 🔒

Enums

CACertificates

ConnectionError

The error type for the MaybeProxyConnector

Constants

ALPN_H2

ALPN identifier for HTTP/2 (RFC 7540 3.1).

BUF_SIZE

Statics

CRYPTO_PROVIDER_CACHE 🔒

RUSTLS_PLATFORM_VERIFIER_CACHE 🔒

A cache for the default rustls platform verifier.

Functions

create_http_client

create_tls_config

Create a TlsConfig to use for managing a HTTP connection. This currently creates a rustls ClientConfig.

prewarm_tls

Prewarm the TLS stack to speed up the first connection

protocol_version_to_string 🔒

Type Aliases

BoxError 🔒

BoxedBody

Connector

ServoClient

TlsConfig