Struct net::device::Device

pub struct Device {
    pub interface: Interface,
    pub name: String,
    pub interface_name: String,
    pub protocols: Vec<Arc<RwLock<Option<Box<dyn Service + Send>>>>>,
    pub ports: Vec<Arc<RwLock<Option<Box<dyn Service + Send>>>>>,
    pub ip_addr: Option<Address>,
    pub dns_addr: Option<Address>,
    pub stream: NetworkStream,
    /* private fields */
}

This is a instance of a device that can be connected to

Fields
interface: Interfacename: Stringinterface_name: Stringprotocols: Vec<Arc<RwLock<Option<Box<dyn Service + Send>>>>>ports: Vec<Arc<RwLock<Option<Box<dyn Service + Send>>>>>ip_addr: Option<Address>dns_addr: Option<Address>stream: NetworkStream
Implementations

impl Device

pub fn start(&mut self)

Start deice for sending/receiving traffic

pub fn set_ip(&mut self, gateway: &str, ip: &str) -> Result<()>

Set gateway and the IP address of this device

Important:

Due to limitations of TUN/TAP, a virtual local area network (VLAN) is created on this system and all virtual network interface cards (VNICs) will will be “plugged into” the vlan. So all devices should have the same gateway but be given different IPs.

Parameter(s):

‘gateway’ - this the gateway to all VNICs on this system (EX: 192.168.69.0/24)

‘ip’ - IP address of this device on the VLAN (EX: 192.168.69.1)

pub fn add_listen_service( &mut self, service: Box<dyn Service + Send>, host_port_num: u16 ) -> Result<()>

Add a new service to listen on a given port number (Event based)

Parameter(s):

‘service’ - the service that will be set to listen on a given port number

‘host_port_num’ - port to use

Error(s):

Returns an error if port is already in use

pub fn stop_listen_service(&mut self, host_port_num: u16)

Stop a service from listening on a given port number

Parameter(s):

‘host_port_num’ - port to stop

pub fn add_listen_service_without_port( &mut self, service: Box<dyn Service + Send>, protocol_num: u16 ) -> Result<()>

Add a new service to listen on a given protocol (Event based) This is intended for services that don’t use port numbers like ICMP

Parameter(s):

‘service’ - the service that will be set to listen on a given port number

‘host_port_num’ - port to use

Error(s):

Returns an error if port is already in use

pub fn stop_listen_service_without_port(&mut self, protocol_num: u16)

Stop a service from listening fro a given protocol number

Parameter(s):

‘host_port_num’ - port to stop