Enum net::IpCidr

pub enum IpCidr {
    Ipv4(Cidr),
    Ipv6(Cidr),
}

A specification of a CIDR block, containing an address and a variable-length subnet masking prefix length.

Variants

Ipv4(Cidr)

Ipv6(Cidr)

Implementations

impl Cidr

pub fn new(addr: Address, prefix_len: u8) -> Cidr

Create a CIDR block from the given address and prefix length.

Panics

This function panics if the given prefix length is invalid for the given address.

pub const fn address(&self) -> Address

Return the IP address of this CIDR block.

pub const fn prefix_len(&self) -> u8

Return the prefix length of this CIDR block.

pub fn contains_addr(&self, addr: &Address) -> bool

Query whether the subnetwork described by this CIDR block contains the given address.

pub fn contains_subnet(&self, subnet: &Cidr) -> bool

Query whether the subnetwork described by this CIDR block contains the subnetwork described by the given CIDR block.