Struct net::fetch::cors_cache::CORSCache

pub struct CORSCache(Vec<CORSCacheEntry>);

A simple, vector-based CORS Cache

Methods

impl CORSCache

fn new() -> CORSCache

fn find_entry_by_header<'a>(&'a mut self, request: &Request, header_name: &str) -> Option<&'a mut CORSCacheEntry>

fn find_entry_by_method<'a>(&'a mut self, request: &Request, method: Method) -> Option<&'a mut CORSCacheEntry>

fn clear(&mut self, request: &Request)

Clear the cache

fn cleanup(&mut self)

Remove old entries

fn match_header(&mut self, request: &Request, header_name: &str) -> bool

Returns true if an entry with a matching header is found

fn match_header_and_update(&mut self, request: &Request, header_name: &str, new_max_age: u32) -> bool

Updates max age if an entry for a matching header is found.

If not, it will insert an equivalent entry

fn match_method(&mut self, request: &Request, method: Method) -> bool

Returns true if an entry with a matching method is found

fn match_method_and_update(&mut self, request: &Request, method: Method, new_max_age: u32) -> bool

Updates max age if an entry for a matching method is found.

If not, it will insert an equivalent entry

fn insert(&mut self, entry: CORSCacheEntry)

Insert an entry