Struct net::fetch::cors_cache::CORSCacheTask

pub struct CORSCacheTask {
    // some fields omitted
}

A simple task-based CORS Cache that can be sent messages

Example

let task = CORSCacheTask::new();
let builder = TaskBuilder::new().named("XHRTask");
let mut sender = task.get_sender();
builder.spawn(move || { task.run() });
sender.insert(CORSCacheEntry::new(/* parameters here */));

Methods

impl CORSCacheTask

fn new() -> CORSCacheTask

fn get_sender(&self) -> CORSCacheSender

Provides a sender to the cache task

fn run(&mut self)

Runs the cache task This blocks the current task, so it is advised to spawn a new task for this Send ExitMsg to the associated Sender to exit