Trait net::fetch::cors_cache::CORSCache

pub trait CORSCache {
    fn clear(&mut self, request: CacheRequestDetails);
    fn cleanup(&mut self);
    fn match_header(&mut self, request: CacheRequestDetails, header_name: &str) -> bool;
    fn match_header_and_update(&mut self, request: CacheRequestDetails, header_name: &str, new_max_age: usize) -> bool;
    fn match_method(&mut self, request: CacheRequestDetails, method: Method) -> bool;
    fn match_method_and_update(&mut self, request: CacheRequestDetails, method: Method, new_max_age: usize) -> bool;
    fn insert(&mut self, entry: CORSCacheEntry);
}

Trait for a generic CORS Cache

Required Methods

fn clear(&mut self, request: CacheRequestDetails)

Clear the cache

fn cleanup(&mut self)

Remove old entries

fn match_header(&mut self, request: CacheRequestDetails, header_name: &str) -> bool

Returns true if an entry with a matching header is found

fn match_header_and_update(&mut self, request: CacheRequestDetails, header_name: &str, new_max_age: usize) -> bool

Updates max age if an entry for a matching header is found.

If not, it will insert an equivalent entry

fn match_method(&mut self, request: CacheRequestDetails, method: Method) -> bool

Returns true if an entry with a matching method is found

fn match_method_and_update(&mut self, request: CacheRequestDetails, method: Method, new_max_age: usize) -> bool

Updates max age if an entry for a matching method is found.

If not, it will insert an equivalent entry

fn insert(&mut self, entry: CORSCacheEntry)

Insert an entry