Struct net::filemanager_thread::FileStoreEntry

struct FileStoreEntry {
    origin: FileOrigin,
    file_impl: FileImpl,
    refs: AtomicUsize,
    is_valid_url: AtomicBool,
}

FileManagerStore's entry

Fields
origin: FileOrigin

Origin of the entry's "creator"

file_impl: FileImpl

Backend implementation

refs: AtomicUsize

Number of FileID holders that the ID is used to index this entry in FileManagerStore. Reference holders include a FileStoreEntry or a script-side File-based Blob

is_valid_url: AtomicBool

UUIDs only become valid blob URIs when explicitly requested by the user with createObjectURL. Validity can be revoked as well. (The UUID is the one that maps to this entry in FileManagerStore)