net

Module http_loader

Structs

HttpState

RedirectEndTimer 🔒

ResponseEndTimer 🔒

Enums

BodyChunk 🔒

Messages from the IPC route to the fetch worker, used to fill the body with bytes coming-in over IPC.

BodySink 🔒

The sink side of the body passed to hyper, used to enqueue chunks.

BodyStream 🔒

The stream side of the body passed to hyper.

CrossOriginResourcePolicy 🔒

https://fetch.spec.whatwg.org/#cross-origin-resource-policy-check

ForNavigation 🔒

HttpCacheEntryState

The various states an entry of the HttpCache can be in.

Constants

FRAGMENT 🔒

Statics

REQUEST_BODY_HEADER_NAMES 🔒

https://fetch.spec.whatwg.org/#request-body-header-name

Functions

append_a_request_origin_header 🔒

https://fetch.spec.whatwg.org/#append-a-request-origin-header

append_cache_data_to_headers 🔒

Steps 8.16 to 8.18 in HTTP network or cache fetch

append_the_fetch_metadata_headers 🔒

https://w3c.github.io/webappsec-fetch-metadata/#abstract-opdef-append-the-fetch-metadata-headers-for-a-request

auth_from_cache 🔒

block_for_cache_ready 🔒

If the cache is not ready to construct a response, wait.

build_tls_security_info 🔒

cors_check 🔒

CORS check

cors_preflight_fetch 🔒

CORS preflight fetch

cross_origin_resource_policy_check 🔒

https://fetch.spec.whatwg.org/#cross-origin-resource-policy-check

cross_origin_resource_policy_internal_check 🔒

https://fetch.spec.whatwg.org/#cross-origin-resource-policy-internal-check

determine_requests_referrer

https://w3c.github.io/webappsec-referrer-policy/#determine-requests-referrer

has_credentials 🔒

http_fetch 🔒

HTTP fetch

http_network_fetch 🔒

HTTP network fetch

http_network_or_cache_fetch 🔒

HTTP network or cache fetch

http_redirect_fetch

HTTP redirect fetch

is_no_store_cache 🔒

is_redirect_status 🔒

https://fetch.spec.whatwg.org/#redirect-status

is_schemelessy_same_site 🔒

https://html.spec.whatwg.org/multipage/#schemelessly-same-site

location_url_for_response 🔒

https://fetch.spec.whatwg.org/#concept-response-location-url

log_fetch_terminated_send_failure 🔒

log_request_body_stream_closed 🔒

no_referrer_when_downgrade 🔒

https://w3c.github.io/webappsec-referrer-policy/#referrer-policy-state-no-referrer-when-downgrade

obtain_response 🔒

This sets up the callback infrastructure to send body frames to body_sender and fires the client request.

obtain_response_setup_router_callback 🔒

Setup the callback mechanism to forward chunks from the request received to the chunk_requester.

origin_when_cross_origin 🔒

https://w3c.github.io/webappsec-referrer-policy/#referrer-policy-origin-when-cross-origin

request_body_stream_closed_error 🔒

same_origin 🔒

https://w3c.github.io/webappsec-referrer-policy/#referrer-policy-same-origin

serialize_origin

Step 3 of https://fetch.spec.whatwg.org/#serializing-a-request-origin.

serialize_request_origin 🔒

https://fetch.spec.whatwg.org/#serializing-a-request-origin

set_cookie_for_url 🔒

set_cookies_from_headers 🔒

set_default_accept 🔒

Step 11 of https://fetch.spec.whatwg.org/#concept-fetch.

set_default_accept_encoding 🔒

set_request_cookies 🔒

set_requests_referrer_policy_on_redirect 🔒

https://w3c.github.io/webappsec-referrer-policy/#set-requests-referrer-policy-on-redirect

set_the_sec_fetch_dest_header 🔒

https://w3c.github.io/webappsec-fetch-metadata/#abstract-opdef-set-dest

set_the_sec_fetch_mode_header 🔒

https://w3c.github.io/webappsec-fetch-metadata/#abstract-opdef-set-mode

set_the_sec_fetch_site_header 🔒

https://w3c.github.io/webappsec-fetch-metadata/#abstract-opdef-set-site

set_the_sec_fetch_user_header 🔒

https://w3c.github.io/webappsec-fetch-metadata/#abstract-opdef-set-user

spawn_stale_while_revalidate 🔒

The cached (stale) response has already been returned to the caller; here we fire off an independent fetch whose only purpose is to refresh the stored response.

strict_origin 🔒

https://w3c.github.io/webappsec-referrer-policy/#referrer-policy-strict-origin

strict_origin_when_cross_origin 🔒

https://w3c.github.io/webappsec-referrer-policy/#referrer-policy-strict-origin-when-cross-origin

strip_url_for_use_as_referrer 🔒

https://w3c.github.io/webappsec-referrer-policy/#strip-url

tao_check 🔒

https://fetch.spec.whatwg.org/#concept-tao-check

wait_for_inflight_requests 🔒

Wait for a cached response from channel. Happens when a fetch gets a cache hit, and the resource is pending completion from the network.