Type Alias net::http_loader::HttpCacheState

type HttpCacheState = Mutex<HashMap<CacheKey, Arc<(Mutex<HttpCacheEntryState>, Condvar)>>>;

Aliased Type

struct HttpCacheState {
    inner: Mutex,
    poison: Flag,
    data: UnsafeCell<HashMap<CacheKey, Arc<(Mutex<HttpCacheEntryState>, Condvar)>, RandomState>>,
}

Fields
inner: Mutexpoison: Flagdata: UnsafeCell<HashMap<CacheKey, Arc<(Mutex<HttpCacheEntryState>, Condvar)>, RandomState>>
Implementations

impl<T> Mutex<T>

pub const fn new(t: T) -> Mutex<T>

Creates a new mutex in an unlocked state ready for use.

Examples

use std::sync::Mutex;

let mutex = Mutex::new(0);

impl<T> Mutex<T>where T: ?Sized,

pub fn lock(&self) -> Result<MutexGuard<'_, T>, PoisonError<MutexGuard<'_, T>>>

Acquires a mutex, blocking the current thread until it is able to do so.

This function will block the local thread until it is available to acquire the mutex. Upon returning, the thread is the only thread with the lock held. An RAII guard is returned to allow scoped unlock of the lock. When the guard goes out of scope, the mutex will be unlocked.

The exact behavior on locking a mutex in the thread which already holds the lock is left unspecified. However, this function will not return on the second call (it might panic or deadlock, for example).

Errors

If another user of this mutex panicked while holding the mutex, then this call will return an error once the mutex is acquired.

Panics

This function might panic when called if the lock is already held by the current thread.

Examples

use std::sync::{Arc, Mutex};
use std::thread;

let mutex = Arc::new(Mutex::new(0));
let c_mutex = Arc::clone(&mutex);

thread::spawn(move || {
    *c_mutex.lock().unwrap() = 10;
}).join().expect("thread::spawn failed");
assert_eq!(*mutex.lock().unwrap(), 10);

pub fn try_lock( &self ) -> Result<MutexGuard<'_, T>, TryLockError<MutexGuard<'_, T>>>

Attempts to acquire this lock.

If the lock could not be acquired at this time, then Err is returned. Otherwise, an RAII guard is returned. The lock will be unlocked when the guard is dropped.

This function does not block.

Errors

If another user of this mutex panicked while holding the mutex, then this call will return the Poisoned error if the mutex would otherwise be acquired.

If the mutex could not be acquired because it is already locked, then this call will return the WouldBlock error.

Examples

use std::sync::{Arc, Mutex};
use std::thread;

let mutex = Arc::new(Mutex::new(0));
let c_mutex = Arc::clone(&mutex);

thread::spawn(move || {
    let mut lock = c_mutex.try_lock();
    if let Ok(ref mut mutex) = lock {
        **mutex = 10;
    } else {
        println!("try_lock failed");
    }
}).join().expect("thread::spawn failed");
assert_eq!(*mutex.lock().unwrap(), 10);

pub fn unlock(guard: MutexGuard<'_, T>)

🔬This is a nightly-only experimental API. (mutex_unlock)

Immediately drops the guard, and consequently unlocks the mutex.

This function is equivalent to calling drop on the guard but is more self-documenting. Alternately, the guard will be automatically dropped when it goes out of scope.

#![feature(mutex_unlock)]

use std::sync::Mutex;
let mutex = Mutex::new(0);

let mut guard = mutex.lock().unwrap();
*guard += 20;
Mutex::unlock(guard);

pub fn is_poisoned(&self) -> bool

Determines whether the mutex is poisoned.

If another thread is active, the mutex can still become poisoned at any time. You should not trust a false value for program correctness without additional synchronization.

Examples

use std::sync::{Arc, Mutex};
use std::thread;

let mutex = Arc::new(Mutex::new(0));
let c_mutex = Arc::clone(&mutex);

let _ = thread::spawn(move || {
    let _lock = c_mutex.lock().unwrap();
    panic!(); // the mutex gets poisoned
}).join();
assert_eq!(mutex.is_poisoned(), true);

pub fn clear_poison(&self)

🔬This is a nightly-only experimental API. (mutex_unpoison)

Clear the poisoned state from a mutex

If the mutex is poisoned, it will remain poisoned until this function is called. This allows recovering from a poisoned state and marking that it has recovered. For example, if the value is overwritten by a known-good value, then the mutex can be marked as un-poisoned. Or possibly, the value could be inspected to determine if it is in a consistent state, and if so the poison is removed.

Examples

#![feature(mutex_unpoison)]

use std::sync::{Arc, Mutex};
use std::thread;

let mutex = Arc::new(Mutex::new(0));
let c_mutex = Arc::clone(&mutex);

let _ = thread::spawn(move || {
    let _lock = c_mutex.lock().unwrap();
    panic!(); // the mutex gets poisoned
}).join();

assert_eq!(mutex.is_poisoned(), true);
let x = mutex.lock().unwrap_or_else(|mut e| {
    **e.get_mut() = 1;
    mutex.clear_poison();
    e.into_inner()
});
assert_eq!(mutex.is_poisoned(), false);
assert_eq!(*x, 1);

pub fn into_inner(self) -> Result<T, PoisonError<T>>

Consumes this mutex, returning the underlying data.

Errors

If another user of this mutex panicked while holding the mutex, then this call will return an error instead.

Examples

use std::sync::Mutex;

let mutex = Mutex::new(0);
assert_eq!(mutex.into_inner().unwrap(), 0);

pub fn get_mut(&mut self) -> Result<&mut T, PoisonError<&mut T>>

Returns a mutable reference to the underlying data.

Since this call borrows the Mutex mutably, no actual locking needs to take place – the mutable borrow statically guarantees no locks exist.

Errors

If another user of this mutex panicked while holding the mutex, then this call will return an error instead.

Examples

use std::sync::Mutex;

let mut mutex = Mutex::new(0);
*mutex.get_mut().unwrap() = 10;
assert_eq!(*mutex.lock().unwrap(), 10);