Struct net::image::holder::ImageHolder

pub struct ImageHolder<NodeAddress> {
    // some fields omitted
}

A struct to store image data. The image will be loaded once the first time it is requested, and an Arc will be stored. Clones of this Arc are given out on demand.

Methods

impl<NodeAddress: Send + 'static> ImageHolder<NodeAddress>

fn new(url: Url, local_image_cache: Arc<Mutex<LocalImageCache<NodeAddress>>>) -> ImageHolder<NodeAddress>

fn size(&self) -> Size2D<isize>

This version doesn't perform any computation, but may be stale w.r.t. newly-available image data that determines size.

The intent is that the impure version is used during layout when dimensions are used for computing layout.

fn get_size(&mut self, node_address: NodeAddress) -> Option<Size2D<isize>>

Query and update the current image size.

fn get_image_if_present(&self) -> Option<Arc<Box<Image>>>

fn get_image(&mut self, node_address: NodeAddress) -> Option<Arc<Box<Image>>>

fn url(&self) -> &Url