net::image_cache

Struct ImageCacheImpl

pub struct ImageCacheImpl {
    store: Arc<Mutex<ImageCacheStore>>,
    svg_id_image_id_map: Arc<Mutex<FxHashMap<Uuid, PendingImageId>>>,
    broken_image_icon_data: Arc<Vec<u8>>,
    thread_pool: Arc<ThreadPool>,
    usvg_options: Arc<Options<'static>>,
    usvg_font_resolver: Arc<dyn FontResolver>,
}

Fields
store: Arc<Mutex<ImageCacheStore>>

Per-ImageCache data.

svg_id_image_id_map: Arc<Mutex<FxHashMap<Uuid, PendingImageId>>>

Maps an SVGElement uuid to a pending image id in the store

broken_image_icon_data: Arc<Vec<u8>>

The data to use for the broken image icon used when images cannot load.

thread_pool: Arc<ThreadPool>

Thread pool for image decoding. This is shared with other ImageCaches in the same process.

usvg_options: Arc<Options<'static>>

The options for usvg. Contains a fontdb::Database and fontresolver.

usvg_font_resolver: Arc<dyn FontResolver>

A font resolve used for resolving fonts when rasterizing SVGs.

This is only used inside usvg::Options but is here so we can measure it.

Implementations

impl ImageCacheImpl

fn add_listener_with_store( &self, store: &mut ImageCacheStore, listener: ImageLoadListener, )

Require self.store.lock() before calling.