Enum net::image_cache_task::Msg

pub enum Msg {
    Prefetch(Url),
    Decode(Url),
    GetImage(Url, Sender<ImageResponseMsg>),
    WaitForImage(Url, Sender<ImageResponseMsg>),
    Exit(Sender<()>),
    StorePrefetchedImageData(Url, Result<Vec<u8>, ()>),
    StoreImage(Url, Option<Arc<Box<Image>>>),
    WaitForStore(Sender<()>),
    WaitForStorePrefetched(Sender<()>),
}

Variants

Prefetch

Tell the cache that we may need a particular image soon. Must be posted before Decode

Decode

Tell the cache to decode an image. Must be posted before GetImage/WaitForImage

GetImage

Request an Image object for a URL. If the image is not is not immediately available then ImageNotReady is returned.

WaitForImage

Wait for an image to become available (or fail to load).

Exit

Clients must wait for a response before shutting down the ResourceTask

StorePrefetchedImageData

Used by the prefetch tasks to post back image binaries

StoreImage

Used by the decoder tasks to post decoded images back to the cache

WaitForStore

For testing

WaitForStorePrefetched

For testing