Module net::image_cache_thread

Structs

AllPendingLoads

CompletedLoad
Represents an image that has completed loading. Images that fail to load (due to network or decode failure) are still stored here, so that they aren't fetched again.

DecoderMsg
Message that the decoder worker threads send to main image cache thread.

ImageCache
Implementation of the image cache

ImageCacheOrigin

ImageListener
Stores information to notify a client when the state of an image changes.

LoadKey

LoadKeyGenerator

PendingLoad
TODO(gw): Remaining work on image cache: * Make use of the prefetch support in various parts of the code. * Profile time in GetImageIfAvailable - might be worth caching these results per paint / layout thread.

Receivers

ResourceLoadInfo

Enums

CacheResult

LoadResult

SelectResult
The types of messages that the main image cache thread receives.

Functions

convert_format

new_image_cache_thread
Create a new image cache.