Struct net::image_cache_thread::ImageCache

struct ImageCache {
    decoder_sender: Sender<DecoderMsg>,
    thread_pool: ThreadPool,
    pending_loads: AllPendingLoads,
    completed_loads: HashMap<ServoUrl, CompletedLoad>,
    placeholder_image: Option<Arc<Image>>,
    webrender_api: RenderApi,
}

Implementation of the image cache

Fields
decoder_sender: Sender<DecoderMsg> thread_pool: ThreadPool pending_loads: AllPendingLoads completed_loads: HashMap<ServoUrl, CompletedLoad> placeholder_image: Option<Arc<Image>> webrender_api: RenderApi
Methods

impl ImageCache

fn run(webrender_api: RenderApi,
ipc_command_receiver: IpcReceiver<ImageCacheCommand>)

fn handle_cmd(&mut self, cmd: ImageCacheCommand) -> Option<IpcSender<()>>

fn handle_progress(&mut self, msg: ResourceLoadInfo)

fn handle_decoder(&mut self, msg: DecoderMsg)

fn complete_load(&mut self, key: PendingImageId, load_result: LoadResult)

fn add_listener(&mut self, id: PendingImageId, listener: ImageResponder)

Add a listener for a given image if it is still pending, or notify the listener if the image is complete.

fn get_completed_image_if_available(&self,
url: &ServoUrl,
placeholder: UsePlaceholder)
-> Option<Result<ImageOrMetadataAvailable, ImageState>>

Return a completed image if it exists, or None if there is no complete load or the complete load is not fully decoded or is unavailable.

fn get_image_or_meta_if_available(&mut self,
url: ServoUrl,
placeholder: UsePlaceholder,
can_request: CanRequestImages)
-> Result<ImageOrMetadataAvailable, ImageState>

Return any available metadata or image for the given URL, or an indication that the image is not yet available if it is in progress, or else reserve a slot in the cache for the URL if the consumer can request images.