Struct net::image_cache_thread::PendingLoad

struct PendingLoad {
    bytes: ImageBytes,
    metadata: Option<ImageMetadata>,
    result: Option<Result<(), NetworkError>>,
    listeners: Vec<ImageResponder>,
    url: ServoUrl,
}

Represents an image that is either being loaded by the resource thread, or decoded by a worker thread.

Fields
bytes: ImageBytes metadata: Option<ImageMetadata> result: Option<Result<(), NetworkError>> listeners: Vec<ImageResponder> url: ServoUrl
Methods

impl PendingLoad

fn new(url: ServoUrl) -> PendingLoad

fn add_listener(&mut self, listener: ImageResponder)