Module net::layer_3::ipv4

Structs

Address

IPv4 IP address object

DifferentiatedServices

ECN

Explicit Congestion Notification (ECN) is a much more recent addition to IPV4 and is not fully supported by all protocols yet has become critical to TCP particularly in recent years. The role of the ECN to act as a “control flow” of data through a medium. This is how systems are able to dynamically scale the bandwidth they occupy. ECN is implemented differently protocol to protocol. In the case of TCP/IP if data is having trouble being delivered, a congestion marker is sent in a SYN packet.

Flags

Flags configure data about fragmentation. If “don’t fragment” and “more fragments” are set, the packet will just be dropped

IPV4

IPV4 datagram

Enums

AssuredForwarding

Assured Forwarding is the second half of the DifferServ segment. These values are ignored if the router is configured for ToS or Ip Precedence only. However, almost all routers today are capable of understanding these. Official Assured Forwarding numbers go as AF11 or AF33 where the first number, 1 in the case of AF13, is the “class” or ip precedence number and the second number is the Assured Forwarding number which are declared in this enum. The number of precedence are backwards compared to IP precedence or classes section as the lower the number, the higher the priority it is in queue

IPPrecedence

IP Precedence is considered “old” and was originally called “Type of Service” (ToS). Modern routers would still be able to understand this by itself but the up to date version is DiffServ which has a second component to it for more customization of priority. It should be noted that legacy routers will not have any trouble handling DiffServ traffic over ToS traffic as the bits that followed ToS that are now used by DiffServe were marked “reserved” and were just ignored by the routers that that use ToS