Enum net::layer_3::ipv4::IPPrecedence

pub enum IPPrecedence {
    DF,
    CS1,
    CS2,
    CS3,
    CS4,
    CS5,
    CS6,
    CS7,
}

IP Precedence is considered “old” and was originally called “Type of Service” (ToS). Modern routers would still be able to understand this by itself but the up to date version is DiffServ which has a second component to it for more customization of priority. It should be noted that legacy routers will not have any trouble handling DiffServ traffic over ToS traffic as the bits that followed ToS that are now used by DiffServe were marked “reserved” and were just ignored by the routers that that use ToS

Variants

DF

“Default” setting - Used for standard traffic.

CS1

Priority - Data that, if given the choice, should have priority. It’s recommended this to be used for the FTP or SMB protocols for file transfers

CS2

Immediate - This is given higher precedence in transmission. It’s recommended to use this level for ssh, ping, or syslog transmissions

CS3

Flash - Higher priority that should be given to RTSP or streaming video

CS4

Flash-override - Named after the term when the President Of The United States deems a message of highest importance and should override all other messages no matter what they are. These should be assigned to video gaming streams or video conferencing streams.

CS5

Critical - Highest priority given to non-networking protocols. It is important that non networking protocols, like NTP, not go past this level. Some routers will block or drop the transmission as it could be deemed malicious if you go past this level. This level should be used for NTP or Peer-to-Peer (PTP) transmissions, like SIP, H.323, etc. It is also important to know that if you give all transmissions this level of precedence, then you will see no speed gain and might see all transmissions that are CS3 and below be dropped

CS6

(NETWORKING ONLY) Internet - A networking protocol only precedence that should only be given to routing protocols like OSPF, BGP, RIP, ETC.

CS7

(NETWORKING ONLY) Networking - Look at yourself in the mirror. Do you really want this level? Is what you’re sending so important that you are willing to risk not being able to send it at all (as routing protocols will be overriden)? Nothing by default has or should have this level or precedence unless you have some custom protocol that will launch weapons of mass destruction or cure cancer in one transmission.