Struct net::layer_4::tcp::TCPv4

pub struct TCPv4 {

    pub ipv4: IPV4,
    pub src_port: u16,
    pub dst_port: u16,
    pub seq_num: u32,
    pub ack_num: u32,
    pub data_offset: u4,
    pub cwr: bool,
    pub ece: bool,
    pub urg: bool,
    pub ack: bool,
    pub psh: bool,
    pub rst: bool,
    pub syn: bool,
    pub fin: bool,
    pub window_size: u16,
    pub options: Vec<u8>,
    pub data: Vec<u8>,
    /* private fields */

}

IPv4 version of the TCP protocol. TCP is an OSI Layer 4 encapsulation that is stateful

Fields
ipv4: IPV4

IPv4 header

src_port: u16

Port number that the packet originated from

dst_port: u16

The port number that the packet is to be sent to

seq_num: u32

Sequence number

ack_num: u32

Acknowledgment number

data_offset: u4

Tells where the data is (as header can have a dynamic length)

cwr: boolece: boolurg: boolack: boolpsh: boolrst: boolsyn: boolfin: boolwindow_size: u16

Length of data the receiver is willing to get in return

options: Vec<u8>

options values

data: Vec<u8>

Data that is encapsulated in the packet

Implementations

impl TCPv4

pub fn new( src_addr: Address, src_port: u16, dest_addr: Address, dst_port: u16, data: &[u8], options: &[u8], flags: &[TcpFlags], seq_num: u32, ack_num: u32, window_size: u16, urgent_pointer: u16 ) -> TCPv4

Create / initialize a tcp packet

pub fn verify(&mut self) -> bool

returns if this packet passes the checksum

pub fn calc_checksum(&mut self) -> u16

This will calculate the checksum for this packet

pub fn update_checksum(&mut self)

update the checksum of this packet (Needs to be done after any updates)

pub fn encode(&mut self, ignore_ipv4: bool) -> Vec<u8>

encode the data into a vector of u8s

pub fn decode(arr: &[u8]) -> Result<TCPv4>

decode an array of u8s into an tcp packet