net::local_directory_listing

Function build_html_directory_listing

pub(crate) async fn build_html_directory_listing(
    url: &Url,
    path: PathBuf,
    directory_contents: ReadDir,
) -> String

Returns an the string of an JavaScript <script> tag calling the setData function with the contents of the given [ReadDir] directory listing.

Arguments

url - the original URL of the request that triggered this directory listing.

path - the full path to the local directory.

directory_contents - a [ReadDir] with the contents of the directory.