Struct net::phy::PcapWriter

pub struct PcapWriter<D, S>where
    D: for<'a> Device<'a>,
    S: PcapSink,{ /* private fields */ }

A packet capture writer device.

Every packet transmitted or received through this device is timestamped and written (in the libpcap format) using the provided sink. Note that writes are fine-grained, and buffering is recommended.

The packet sink should be cheaply cloneable, as it is cloned on every transmitted packet. For example, &'a mut Vec<u8> is cheaply cloneable but &std::io::File

Implementations

impl<D, S> PcapWriter<D, S>where
D: for<'a> Device<'a>,
S: PcapSink,

pub fn new(lower: D, sink: S, mode: PcapMode) -> PcapWriter<D, S>

Creates a packet capture writer.

pub fn get_ref(&self) -> &D

Get a reference to the underlying device.

Even if the device offers reading through a standard reference, it is inadvisable to directly read from the device as doing so will circumvent the packet capture.

pub fn get_mut(&mut self) -> &mut D

Get a mutable reference to the underlying device.

It is inadvisable to directly read from the device as doing so will circumvent the packet capture.