Struct net::phy::Tracer

pub struct Tracer<D>where
    D: Device,{ /* private fields */ }

A tracer device.

A tracer is a device that pretty prints all packets traversing it using the provided writer function, and then passes them to another device.

Implementations

impl<D> Tracer<D>where
D: Device,

pub fn new(inner: D, writer: fn(_: Instant, _: Packet<'_>)) -> Tracer<D>

Create a tracer device.

pub fn get_ref(&self) -> &D

Get a reference to the underlying device.

Even if the device offers reading through a standard reference, it is inadvisable to directly read from the device as doing so will circumvent the tracing.

pub fn get_mut(&mut self) -> &mut D

Get a mutable reference to the underlying device.

It is inadvisable to directly read from the device as doing so will circumvent the tracing.

pub fn into_inner(self) -> D

Return the underlying device, consuming the tracer.