Trait net::phy::Device

pub trait Device<'a> {
    type RxToken: 'a + RxToken;
    type TxToken: 'a + TxToken;

    fn receive(&'a mut self) -> Option<(Self::RxToken, Self::TxToken)>;
    fn transmit(&'a mut self) -> Option<Self::TxToken>;
    fn capabilities(&self) -> DeviceCapabilities;
}

An interface for sending and receiving raw network frames.

The interface is based on tokens, which are types that allow to receive/transmit a single packet. The receive and transmit functions only construct such tokens, the real sending/receiving operation are performed when the tokens are consumed.

Required Associated Types

type RxToken: 'a + RxToken

type TxToken: 'a + TxToken

Required Methods

fn receive(&'a mut self) -> Option<(Self::RxToken, Self::TxToken)>

Construct a token pair consisting of one receive token and one transmit token.

The additional transmit token makes it possible to generate a reply packet based on the contents of the received packet. For example, this makes it possible to handle arbitrarily large ICMP echo (“ping”) requests, where the all received bytes need to be sent back, without heap allocation.

fn transmit(&'a mut self) -> Option<Self::TxToken>

Construct a transmit token.

fn capabilities(&self) -> DeviceCapabilities

Get a description of device capabilities.