Module net::pub_domains

Implementation of public domain matching.

The list is a file located on the resources folder and loaded once on first need.

The list can be updated with ./mach update-pub-domains from this source: https://publicsuffix.org/list/

This implementation is not strictly following the specification of the list. Wildcards are not restricted to appear only in the leftmost position, but the current list has no such cases so we don't need to make the code more complex for it. The mach update command makes sure that those cases are not present.

Structs

PUB_DOMAINS

Statics

PUB_DOMAINS

Functions

is_pub_domain
Match the given domain against a static list of known public domains

load_pub_domains