Struct net::resource_task::LoadData

pub struct LoadData {
    pub url: Url,
    pub method: Method,
    pub headers: Headers,
    pub preserved_headers: Headers,
    pub data: Option<Vec<u8>>,
    pub cors: Option<ResourceCORSData>,
    pub consumer: Sender<LoadResponse>,
}

Fields

url

method

headers

Headers that will apply to the initial request only

preserved_headers

Headers that will apply to the initial request and any redirects

data

cors

consumer

Methods

impl LoadData

fn new(url: Url, consumer: Sender<LoadResponse>) -> LoadData