Struct net::resource_task::Metadata

pub struct Metadata {
    pub final_url: Url,
    pub content_type: Option<(String, String)>,
    pub charset: Option<String>,
    pub headers: Option<Headers>,
    pub status: Option<RawStatus>,
}

Metadata about a loaded resource, such as is obtained from HTTP headers.

Fields

final_url

Final URL after redirects.

content_type

MIME type / subtype.

charset

Character set.

headers

Headers

status

HTTP Status

Methods

impl Metadata

fn default(url: Url) -> Metadata

Metadata with defaults for everything optional.

fn set_content_type(&mut self, content_type: Option<&Mime>)

Extract the parts of a Mime that we care about.