Module net::resource_thread

A thread that takes a URL and streams back the binary data.

Structs

AuthCache

AuthCacheEntry

CancellableResource
The optional resources required by the CancellationListener

CancellationListener
A listener which is basically a wrapped optional receiver which looks for the load cancellation message. Some of the loading processes always keep an eye out for this message and stop loading stuff once they receive it.

ResourceChannelManager

ResourceManager

Enums

ProgressSender

Functions

new_resource_thread
Create a ResourceThread

read_json_from_file

send_error

start_sending_opt
For use by loaders in responding to a Load message. It takes an optional NetworkError, so that we can extract the SSL Validation errors and take it to the HTML parser

start_sending_sniffed
For use by loaders in responding to a Load message that allows content sniffing.

start_sending_sniffed_opt
For use by loaders in responding to a Load message that allows content sniffing.

write_json_to_file