Struct net::resource_thread::CancellableResource

pub struct CancellableResource {
    cancel_receiver: Receiver<()>,
    resource_id: ResourceId,
    resource_thread: CoreResourceThread,
}

The optional resources required by the CancellationListener

Fields
cancel_receiver: Receiver<()>

The receiver which receives a message on load cancellation

resource_id: ResourceId

The CancellationListener is unique to this ResourceId

resource_thread: CoreResourceThread

If we haven't initiated any cancel requests, then the loaders ask the listener to remove the ResourceId in the HashMap of CoreResourceManager once they finish loading

Methods

impl CancellableResource

fn new(receiver: Receiver<()>, res_id: ResourceId, res_thread: CoreResourceThread) -> CancellableResource