Module net::socket

Communication between endpoints.

The socket module deals with network endpoints and buffering. It provides interfaces for accessing buffers of data, and protocol state machines for filling and emptying these buffers.

The programming interface implemented here differs greatly from the common Berkeley socket interface. Specifically, in the Berkeley interface the buffering is implicit: the operating system decides on the good size for a buffer and manages it. The interface implemented by this module uses explicit buffering: you decide on the good size for a buffer, allocate it, and let the networking stack use it.

Modules

icmp

raw

tcp

udp

Enums

Socket

A network socket.

Traits

AnySocket

A conversion trait for network sockets.