Enum net::storage_task::StorageTaskMsg

pub enum StorageTaskMsg {
    Length(Sender<u32>, Url, StorageType),
    Key(Sender<Option<DOMString>>, Url, StorageType, u32),
    GetItem(Sender<Option<DOMString>>, Url, StorageType, DOMString),
    SetItem(Sender<(bool, Option<DOMString>)>, Url, StorageType, DOMString, DOMString),
    RemoveItem(Sender<Option<DOMString>>, Url, StorageType, DOMString),
    Clear(Sender<bool>, Url, StorageType),
    Exit,
}

Request operations on the storage data associated with a particular url

Variants

Length

gets the number of key/value pairs present in the associated storage data

Key

gets the name of the key at the specified index in the associated storage data

GetItem

gets the value associated with the given key in the associated storage data

SetItem

sets the value of the given key in the associated storage data TODO throw QuotaExceededError in case of error

RemoveItem

removes the key/value pair for the given key in the associated storage data

Clear

clears the associated storage data by removing all the key/value pairs

Exit

shut down this task