Struct net::storage_thread::StorageManager

struct StorageManager {
    port: IpcReceiver<StorageThreadMsg>,
    session_data: HashMap<String, (usize, BTreeMap<String, String>)>,
    local_data: HashMap<String, (usize, BTreeMap<String, String>)>,
    config_dir: Option<PathBuf>,
}

Fields
port: IpcReceiver<StorageThreadMsg> session_data: HashMap<String, (usize, BTreeMap<String, String>)> local_data: HashMap<String, (usize, BTreeMap<String, String>)> config_dir: Option<PathBuf>
Methods

impl StorageManager

fn new(port: IpcReceiver<StorageThreadMsg>,
config_dir: Option<PathBuf>)
-> StorageManager

impl StorageManager

fn start(&mut self)

fn select_data(&self,
storage_type: StorageType)
-> &HashMap<String, (usize, BTreeMap<String, String>)>

fn select_data_mut(&mut self,
storage_type: StorageType)
-> &mut HashMap<String, (usize, BTreeMap<String, String>)>

fn length(&self,
sender: IpcSender<usize>,
url: ServoUrl,
storage_type: StorageType)

fn key(&self,
sender: IpcSender<Option<String>>,
url: ServoUrl,
storage_type: StorageType,
index: u32)

fn keys(&self,
sender: IpcSender<Vec<String>>,
url: ServoUrl,
storage_type: StorageType)

fn set_item(&mut self,
sender: IpcSender<Result<(bool, Option<String>), ()>>,
url: ServoUrl,
storage_type: StorageType,
name: String,
value: String)

Sends Ok(changed, Some(old_value)) in case there was a previous value with the same key name but with different value name otherwise sends Err(()) to indicate that the operation would result in exceeding the quota limit

fn request_item(&self,
sender: IpcSender<Option<String>>,
url: ServoUrl,
storage_type: StorageType,
name: String)

fn remove_item(&mut self,
sender: IpcSender<Option<String>>,
url: ServoUrl,
storage_type: StorageType,
name: String)

Sends Some(old_value) in case there was a previous value with the key name, otherwise sends None

fn clear(&mut self,
sender: IpcSender<bool>,
url: ServoUrl,
storage_type: StorageType)

fn origin_as_string(&self, url: ServoUrl) -> String