Struct net::DeviceCapabilities

#[non_exhaustive]
pub struct DeviceCapabilities {
    pub medium: Medium,
    pub max_transmission_unit: usize,
    pub max_burst_size: Option<usize>,
    pub checksum: ChecksumCapabilities,
}

A description of device capabilities.

Higher-level protocols may achieve higher throughput or lower latency if they consider the bandwidth or packet size limitations.

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

medium: Medium

Medium of the device.

This indicates what kind of packet the sent/received bytes are, and determines some behaviors of Interface. For example, ARP/NDISC address resolution is only done for Ethernet mediums.

max_transmission_unit: usize

Maximum transmission unit.

The network device is unable to send or receive frames larger than the value returned by this function.

For Ethernet devices, this is the maximum Ethernet frame size, including the Ethernet header (14 octets), but not including the Ethernet FCS (4 octets). Therefore, Ethernet MTU = IP MTU + 14.

Note that in Linux and other OSes, “MTU” is the IP MTU, not the Ethernet MTU, even for Ethernet devices. This is a common source of confusion.

Most common IP MTU is 1500. Minimum is 576 (for IPv4) or 1280 (for IPv6). Maximum is 9216 octets.

max_burst_size: Option<usize>

Maximum burst size, in terms of MTU.

The network device is unable to send or receive bursts large than the value returned by this function.

If None, there is no fixed limit on burst size, e.g. if network buffers are dynamically allocated.

checksum: ChecksumCapabilities

Checksum behavior.

If the network device is capable of verifying or computing checksums for some protocols, it can request that the stack not do so in software to improve performance.

Implementations

impl DeviceCapabilities

pub fn ip_mtu(&self) -> usize