net

Trait Network

pub trait Network: Send + Sync {
    // Required methods
    fn id(&self) -> usize;
    fn send(&self, to: usize, data: &[u8]) -> Result<()>;
    fn recv(&self, from: usize) -> Result<Vec<u8>>;
    fn get_connection_stats(&self) -> ConnectionStats;

    // Provided methods
    fn send_buffered(&self, to: usize, data: &[u8]) -> Result<()> { ... }
    fn flush_to(&self, _to: usize) -> Result<()> { ... }
    fn flush_all(&self) -> Result<()> { ... }
}

A MPC network that can be used to send and receive data to and from other parties

Can be used to send to multiple parties in parallel, but sending to the same party must happen in sequence.

Required Methods

fn id(&self) -> usize

The id of the party

fn send(&self, to: usize, data: &[u8]) -> Result<()>

Send data to other party

fn recv(&self, from: usize) -> Result<Vec<u8>>

Receive data from other party

fn get_connection_stats(&self) -> ConnectionStats

Get connection statistics for the Network. The returned HashMap maps party_id to a tuple of (sent_bytes, received_bytes).

Provided Methods

fn send_buffered(&self, to: usize, data: &[u8]) -> Result<()>

Write data to the send buffer without forcing immediate transmission.

The data will be delivered when [flush_to] or [flush_all] is called, or when the internal buffer is full. Implementations without userspace buffering (like crate::tcp::TcpNetwork) simply delegate to [send].

Default: delegates to [send].

fn flush_to(&self, _to: usize) -> Result<()>

Flush any buffered data for the connection to party to.

After this call, all data previously written via [send_buffered] to party to is guaranteed to have been pushed to the kernel.

Default: no-op (for unbuffered implementations).

fn flush_all(&self) -> Result<()>

Flush buffered data for all connections.

Default: no-op.