Struct net::udp::Socket

pub struct Socket<'a> { /* private fields */ }

A User Datagram Protocol socket.

A UDP socket is bound to a specific endpoint, and owns transmit and receive packet buffers.

Implementations

impl<'a> Socket<'a>

pub fn new(
rx_buffer: PacketBuffer<'a, Endpoint>,
tx_buffer: PacketBuffer<'a, Endpoint>
) -> Socket<'a>

Create an UDP socket with the given buffers.

pub fn endpoint(&self) -> ListenEndpoint

Return the bound endpoint.

pub fn hop_limit(&self) -> Option<u8>

Return the time-to-live (IPv4) or hop limit (IPv6) value used in outgoing packets.

See also the set_hop_limit method

pub fn set_hop_limit(&mut self, hop_limit: Option<u8>)

Set the time-to-live (IPv4) or hop limit (IPv6) value used in outgoing packets.

A socket without an explicitly set hop limit value uses the default IANA recommended value (64).

Panics

This function panics if a hop limit value of 0 is given. See RFC 1122  3.2.1.7.

pub fn bind<T>(&mut self, endpoint: T) -> Result<(), BindError>where
T: Into<ListenEndpoint>,

Bind the socket to the given endpoint.

This function returns Err(Error::Illegal) if the socket was open (see is_open), and Err(Error::Unaddressable) if the port in the given endpoint is zero.

pub fn close(&mut self)

Close the socket.

pub fn is_open(&self) -> bool

Check whether the socket is open.

pub fn can_send(&self) -> bool

Check whether the transmit buffer is full.

pub fn can_recv(&self) -> bool

Check whether the receive buffer is not empty.

pub fn packet_recv_capacity(&self) -> usize

Return the maximum number packets the socket can receive.

pub fn packet_send_capacity(&self) -> usize

Return the maximum number packets the socket can transmit.

pub fn payload_recv_capacity(&self) -> usize

Return the maximum number of bytes inside the recv buffer.

pub fn payload_send_capacity(&self) -> usize

Return the maximum number of bytes inside the transmit buffer.

pub fn send(
&mut self,
size: usize,
remote_endpoint: Endpoint
) -> Result<&mut [u8], SendError>

Enqueue a packet to be sent to a given remote endpoint, and return a pointer to its payload.

This function returns Err(Error::Exhausted) if the transmit buffer is full, Err(Error::Unaddressable) if local or remote port, or remote address are unspecified, and Err(Error::Truncated) if there is not enough transmit buffer capacity to ever send this packet.

pub fn send_with<F>(
&mut self,
max_size: usize,
remote_endpoint: Endpoint,
f: F
) -> Result<usize, SendError>where
F: FnOnce(&mut [u8]) -> usize,

Enqueue a packet to be send to a given remote endpoint and pass the buffer to the provided closure. The closure then returns the size of the data written into the buffer.

Also see send.

pub fn send_slice(
&mut self,
data: &[u8],
remote_endpoint: Endpoint
) -> Result<(), SendError>

Enqueue a packet to be sent to a given remote endpoint, and fill it from a slice.

See also send.

pub fn recv(&mut self) -> Result<(&[u8], Endpoint), RecvError>

Dequeue a packet received from a remote endpoint, and return the endpoint as well as a pointer to the payload.

This function returns Err(Error::Exhausted) if the receive buffer is empty.

pub fn recv_slice(
&mut self,
data: &mut [u8]
) -> Result<(usize, Endpoint), RecvError>

Dequeue a packet received from a remote endpoint, copy the payload into the given slice, and return the amount of octets copied as well as the endpoint.

See also recv.

pub fn peek(&mut self) -> Result<(&[u8], &Endpoint), RecvError>

Peek at a packet received from a remote endpoint, and return the endpoint as well as a pointer to the payload without removing the packet from the receive buffer. This function otherwise behaves identically to recv.

It returns Err(Error::Exhausted) if the receive buffer is empty.

pub fn peek_slice(
&mut self,
data: &mut [u8]
) -> Result<(usize, &Endpoint), RecvError>

Peek at a packet received from a remote endpoint, copy the payload into the given slice, and return the amount of octets copied as well as the endpoint without removing the packet from the receive buffer. This function otherwise behaves identically to recv_slice.

See also peek.