Struct net::wire::Ipv6ExtHeader

pub struct Ipv6ExtHeader<T>where
    T: AsRef<[u8]>,{ /* private fields */ }

A read/write wrapper around an IPv6 Extension Header buffer.

Implementations

impl<T> Header<T>where T: AsRef<[u8]>,

Core getter methods relevant to any IPv6 extension header.

pub const fn new_unchecked(buffer: T) -> Header<T>

Create a raw octet buffer with an IPv6 Extension Header structure.

pub fn new_checked(buffer: T) -> Result<Header<T>, Error>

Shorthand for a combination of new_unchecked and check_len.

pub fn check_len(&self) -> Result<(), Error>

Ensure that no accessor method will panic if called. Returns Err(Error) if the buffer is too short.

The result of this check is invalidated by calling set_header_len.

pub fn into_inner(self) -> T

Consume the header, returning the underlying buffer.

pub fn next_header(&self) -> Protocol

Return the next header field.

pub fn header_len(&self) -> u8

Return the header length field.

impl<'h, T> Header<&'h T>where T: AsRef<[u8]> + ?Sized,

pub fn payload(&self) -> &'h [u8]

Return the payload of the IPv6 extension header.

impl<T> Header<T>where T: AsRef<[u8]> + AsMut<[u8]>,

pub fn set_next_header(&mut self, value: Protocol)

Set the next header field.

pub fn set_header_len(&mut self, value: u8)

Set the extension header data length. The length of the header is in 8-octet units, not including the first 8 octets.

impl<'a, T> Header<&'a mut T>where T: AsRef<[u8]> + AsMut<[u8]> + ?Sized,

pub fn payload_mut(&mut self) -> &mut [u8]

Return a mutable pointer to the payload data.