Struct net::wire::TcpPacket

pub struct TcpPacket<T>where
    T: AsRef<[u8]>,{ /* private fields */ }

A read/write wrapper around a Transmission Control Protocol packet buffer.

Implementations

impl<T> Packet<T>where T: AsRef<[u8]>,

pub const fn new_unchecked(buffer: T) -> Packet<T>

Imbue a raw octet buffer with TCP packet structure.

pub fn new_checked(buffer: T) -> Result<Packet<T>, Error>

Shorthand for a combination of new_unchecked and check_len.

pub fn check_len(&self) -> Result<(), Error>

Ensure that no accessor method will panic if called. Returns Err(Error) if the buffer is too short. Returns Err(Error) if the header length field has a value smaller than the minimal header length.

The result of this check is invalidated by calling set_header_len.

pub fn into_inner(self) -> T

Consume the packet, returning the underlying buffer.

pub fn src_port(&self) -> u16

Return the source port field.

pub fn dst_port(&self) -> u16

Return the destination port field.

pub fn seq_number(&self) -> SeqNumber

Return the sequence number field.

pub fn ack_number(&self) -> SeqNumber

Return the acknowledgement number field.

pub fn fin(&self) -> bool

Return the FIN flag.

pub fn syn(&self) -> bool

Return the SYN flag.

pub fn rst(&self) -> bool

Return the RST flag.

pub fn psh(&self) -> bool

Return the PSH flag.

pub fn ack(&self) -> bool

Return the ACK flag.

pub fn urg(&self) -> bool

Return the URG flag.

pub fn ece(&self) -> bool

Return the ECE flag.

pub fn cwr(&self) -> bool

Return the CWR flag.

pub fn ns(&self) -> bool

Return the NS flag.

pub fn header_len(&self) -> u8

Return the header length, in octets.

pub fn window_len(&self) -> u16

Return the window size field.

pub fn checksum(&self) -> u16

Return the checksum field.

pub fn urgent_at(&self) -> u16

Return the urgent pointer field.

pub fn segment_len(&self) -> usize

Return the length of the segment, in terms of sequence space.

pub fn selective_ack_permitted(&self) -> Result<bool, Error>

Returns whether the selective acknowledgement SYN flag is set or not.

pub fn selective_ack_ranges(&self) -> Result<[Option<(u32, u32)>; 3], Error>

Return the selective acknowledgement ranges, if any. If there are none in the packet, an array of None values will be returned.

pub fn verify_checksum(&self, src_addr: &Address, dst_addr: &Address) -> bool

Validate the packet checksum.

Panics

This function panics unless src_addr and dst_addr belong to the same family, and that family is IPv4 or IPv6.

Fuzzing

This function always returns true when fuzzing.

impl<'a, T> Packet<&'a T>where T: AsRef<[u8]> + ?Sized,

pub fn options(&self) -> &'a [u8]

Return a pointer to the options.

pub fn payload(&self) -> &'a [u8]

Return a pointer to the payload.

impl<T> Packet<T>where T: AsRef<[u8]> + AsMut<[u8]>,

pub fn set_src_port(&mut self, value: u16)

Set the source port field.

pub fn set_dst_port(&mut self, value: u16)

Set the destination port field.

pub fn set_seq_number(&mut self, value: SeqNumber)

Set the sequence number field.

pub fn set_ack_number(&mut self, value: SeqNumber)

Set the acknowledgement number field.

pub fn clear_flags(&mut self)

Clear the entire flags field.

pub fn set_fin(&mut self, value: bool)

Set the FIN flag.

pub fn set_syn(&mut self, value: bool)

Set the SYN flag.

pub fn set_rst(&mut self, value: bool)

Set the RST flag.

pub fn set_psh(&mut self, value: bool)

Set the PSH flag.

pub fn set_ack(&mut self, value: bool)

Set the ACK flag.

pub fn set_urg(&mut self, value: bool)

Set the URG flag.

pub fn set_ece(&mut self, value: bool)

Set the ECE flag.

pub fn set_cwr(&mut self, value: bool)

Set the CWR flag.

pub fn set_ns(&mut self, value: bool)

Set the NS flag.

pub fn set_header_len(&mut self, value: u8)

Set the header length, in octets.

pub fn set_window_len(&mut self, value: u16)

Return the window size field.

pub fn set_checksum(&mut self, value: u16)

Set the checksum field.

pub fn set_urgent_at(&mut self, value: u16)

Set the urgent pointer field.

pub fn fill_checksum(&mut self, src_addr: &Address, dst_addr: &Address)

Compute and fill in the header checksum.

Panics

This function panics unless src_addr and dst_addr belong to the same family, and that family is IPv4 or IPv6.

pub fn options_mut(&mut self) -> &mut [u8]

Return a pointer to the options.

pub fn payload_mut(&mut self) -> &mut [u8]

Return a mutable pointer to the payload data.