Crate net

Modules

about_loader

cookie
Implementation of cookie creation and matching as specified by http://tools.ietf.org/html/rfc6265

cookie_storage
Implementation of cookie storage as specified in http://tools.ietf.org/html/rfc6265

data_loader

fetch
An implementation of the Fetch spec

file_loader

http_loader

image
Image handling.

image_cache_task

local_image_cache
An adapter for ImageCacheTask that does local caching to avoid extra message traffic, it also avoids waiting on the same image multiple times and thus triggering reflows multiple times.

pub_domains

resource_task
A task that takes a URL and streams back the binary data.

storage_task