Crate net

Modules

connector

cookie

Implementation of cookie creation and matching as specified by http://tools.ietf.org/html/rfc6265

cookie_storage

Implementation of cookie storage as specified in http://tools.ietf.org/html/rfc6265

data_loader

decoder

Adapted from an implementation in reqwest. A potentially non-blocking response decoder.

fetch

An implementation of the Fetch specification

filemanager_thread

hosts

hsts

http_cache

A memory cache implementing the logic specified in http://tools.ietf.org/html/rfc7234 and http://tools.ietf.org/html/rfc7232.

http_loader

image_cache

mime_classifier

resource_thread

A thread that takes a URL and streams back the binary data.

storage_thread

subresource_integrity

test

A module for re-exports of items used in unit tests.

websocket_loader

The websocket handler has three main responsibilities: