Crate net

Modules

icmp

phy

Access to networking hardware.

tcp

udp

wire

Low-level packet access and construction.

Structs

DeviceCapabilities

A description of device capabilities.

Instant

A representation of an absolute time value.

IpEndpoint

An internet endpoint address.

LockedSocket

NetworkInterface

A network interface.

Socket

A network socket.

SocketSet

An extensible set of sockets.

Enums

Error

IpAddress

An internetworking address.

IpCidr

A specification of a CIDR block, containing an address and a variable-length subnet masking prefix length.

Traits

NetworkDevice

A network device.

Functions

get_default_interface

Returns the first available interface.

get_ephemeral_port

Returns a port in the range reserved for private, dynamic, and ephemeral ports.

get_interfaces

Returns a list of available interfaces behind a mutex.

register_device

Registers a network device.

Type Definitions

Result