Type Alias net::connector::TlsConfig

pub type TlsConfig = ClientConfig;

Aliased Type

struct TlsConfig {
    pub(super) cipher_suites: Vec<SupportedCipherSuite, Global>,
    pub(super) kx_groups: Vec<&'static SupportedKxGroup, Global>,
    pub alpn_protocols: Vec<Vec<u8, Global>, Global>,
    pub resumption: Resumption,
    pub max_fragment_size: Option<usize>,
    pub client_auth_cert_resolver: Arc<dyn ResolvesClientCert, Global>,
    pub(super) versions: EnabledVersions,
    pub enable_sni: bool,
    pub(super) verifier: Arc<dyn ServerCertVerifier, Global>,
    pub key_log: Arc<dyn KeyLog, Global>,
    pub enable_early_data: bool,
}

Fields
cipher_suites: Vec<SupportedCipherSuite, Global>kx_groups: Vec<&'static SupportedKxGroup, Global>alpn_protocols: Vec<Vec<u8, Global>, Global>

Which ALPN protocols we include in our client hello. If empty, no ALPN extension is sent.

resumption: Resumption

How and when the client can resume a previous session.

max_fragment_size: Option<usize>

The maximum size of TLS message we’ll emit. If None, we don’t limit TLS message lengths except to the 2**16 limit specified in the standard.

rustls enforces an arbitrary minimum of 32 bytes for this field. Out of range values are reported as errors from ClientConnection::new.

Setting this value to the TCP MSS may improve latency for stream-y workloads.

client_auth_cert_resolver: Arc<dyn ResolvesClientCert, Global>

How to decide what client auth certificate/keys to use.

versions: EnabledVersionsenable_sni: bool

Whether to send the Server Name Indication (SNI) extension during the client handshake.

The default is true.

verifier: Arc<dyn ServerCertVerifier, Global>key_log: Arc<dyn KeyLog, Global>

How to output key material for debugging. The default does nothing.

enable_early_data: bool

Whether to send data on the first flight (“early data”) in TLS 1.3 handshakes.

The default is false.