Module net::fetch

An implementation of the Fetch specification

Modules

cors_cache

An implementation of the CORS preflight cache For now this library is XHR-specific. For stuff involving <img>, <iframe>, <form>, etc please check what the request mode should be and compare with the fetch spec This library will eventually become the core of the Fetch crate with CORSRequest being expanded into FetchRequest (etc)

headers

methods