Module net::fetch::cors_cache

An implementation of the CORS preflight cache For now this library is XHR-specific. For stuff involving <img>, <iframe>, <form>, etc please check what the request mode should be and compare with the fetch spec This library will eventually become the core of the Fetch crate with CORSRequest being expanded into FetchRequest (etc)

Structs

BasicCORSCache
A simple, vector-based CORS Cache

CORSCacheEntry
An entry in the CORS cache

CORSCacheTask
A simple task-based CORS Cache that can be sent messages

CacheRequestDetails
Properties of Request required to cache match.

Enums

CORSCacheTaskMsg
Various messages that can be sent to a CORSCacheTask

HeaderOrMethod
Union type for CORS cache entries

Traits

CORSCache
Trait for a generic CORS Cache

Type Definitions

CORSCacheSender
A Sender to a CORSCacheTask