Module net::fetch::cors_cache

An implementation of the CORS preflight cache For now this library is XHR-specific. For stuff involving <img>, <iframe>, <form>, etc please check what the request mode should be and compare with the fetch spec This library will eventually become the core of the Fetch crate with CORSRequest being expanded into FetchRequest (etc)

Structs

CorsCache

A simple, vector-based CORS Cache

CorsCacheEntry

An entry in the CORS cache

Enums

HeaderOrMethod

Union type for CORS cache entries

Functions

match_headers