Module net::fetch::methods

Structs

CancellationListener

FetchContext

X_CONTENT_TYPE_OPTIONS 🔒

Enums

Data

RangeRequestBounds

Range header start and end values.

Functions

create_blank_reply 🔒

fetch

Fetch

fetch_with_cors_cache

get_range_request_bounds 🔒

Get the range bounds if the Range header is present.

handle_allowcert_request 🔒

Handle a request from the user interface to ignore validation errors for a certificate.

is_bad_port 🔒

https://fetch.spec.whatwg.org/#bad-port

is_network_scheme 🔒

https://fetch.spec.whatwg.org/#network-scheme

is_null_body_status 🔒

main_fetch

Main fetch

partial_content 🔒

range_not_satisfiable_error 🔒

scheme_fetch 🔒

Scheme fetch

should_be_blocked_due_to_bad_port

https://fetch.spec.whatwg.org/#block-bad-port

should_be_blocked_due_to_mime_type 🔒

https://fetch.spec.whatwg.org/#should-response-to-request-be-blocked-due-to-mime-type?

should_be_blocked_due_to_nosniff

https://fetch.spec.whatwg.org/#should-response-to-request-be-blocked-due-to-nosniff?

should_request_be_blocked_by_csp

https://www.w3.org/TR/CSP/#should-block-request

wait_for_response 🔒

Type Aliases

DoneChannel

Target