net

Module hsts

Structs

HstsEntry

HstsList

HstsPreloadList

Represents the portion of the HSTS list that comes from the preload list it is split out to allow sharing between the private and public http state as well as potentially swpaping out the underlying type to something immutable and more efficient like FSTs or DAFSA/DAWGs. To generate a new version of the FST map file run ./mach update-hsts-preload

Statics

PRELOAD_LIST_ENTRIES 🔒

Functions

hsts_preload_size_of

unix_timestamp_to_nonzerou64 🔒