net

Module http_cache

A memory cache implementing the logic specified in http://tools.ietf.org/html/rfc7234 and http://tools.ietf.org/html/rfc7232.

Structs

CacheKey

The key used to differentiate requests in the cache.

CachedMetadata 🔒

Metadata about a loaded resource, such as is obtained from HTTP headers.

CachedResource

A complete cached resource.

CachedResponse 🔒

Wrapper around a cached response, including information on re-validation needs

HttpCache

A simple memory cache. Elements will be evicted based on the cache heuristic. We weight elements by the number of entries per given url. We evict currently a whole url. The cache makes extensive use of Arc::unwrap_or_clone or and Arc::into_inner to modify the cached entries. This is ok because CachedResource are cheap to clone

Enums

CachedResourcesOrGuard

Returns an writeable entry into the cache or a guard for insertint an entry The guard will block other queries to the cache entry in both cases.

Functions

calculate_response_age 🔒

Calculating Age https://tools.ietf.org/html/rfc7234#section-4.2.3

construct_response 🔒

Constructing Responses from Caches. https://tools.ietf.org/html/rfc7234#section-4

create_cached_response 🔒

Create a CachedResponse from a request and a CachedResource.

create_resource_with_bytes_from_resource 🔒

Create a new resource, based on the bytes requested, and an existing resource, with a status-code of 206.

get_expiry_adjustment_from_request_headers 🔒

Request Cache-Control Directives https://tools.ietf.org/html/rfc7234#section-5.2.1

get_response_expiry 🔒

Determine the expiry date from relevant headers, or uses a heuristic if none are present.

handle_range_request 🔒

Support for range requests https://tools.ietf.org/html/rfc7233.

invalidate 🔒

Invalidation. https://tools.ietf.org/html/rfc7234#section-4.4

invalidate_cached_resources 🔒

is_cacheable_by_default 🔒

Determine if a response is cacheable by default https://tools.ietf.org/html/rfc7231#section-6.1

refresh

Freshening Stored Responses upon Validation. https://tools.ietf.org/html/rfc7234#section-4.3.4

resolve_location_url 🔒

response_is_cacheable 🔒

Determine if a given response is cacheable. Based on https://tools.ietf.org/html/rfc7234#section-3

Type Aliases

CacheEntry 🔒

OurLifecycle 🔒

QuickCache 🔒

QuickCachePlaceeholderGuard 🔒