net

Module http_loader

Structs

HttpState

RedirectEndTimer ğŸ”’

ResponseEndTimer ğŸ”’

Enums

BodyChunk ğŸ”’

Messages from the IPC route to the fetch worker, used to fill the body with bytes coming-in over IPC.

BodySink ğŸ”’

The sink side of the body passed to hyper, used to enqueue chunks.

BodyStream ğŸ”’

The stream side of the body passed to hyper.

CrossOriginResourcePolicy ğŸ”’

https://fetch.spec.whatwg.org/#cross-origin-resource-policy-check

HttpCacheEntryState

The various states an entry of the HttpCache can be in.

Statics

REQUEST_BODY_HEADER_NAMES ğŸ”’

https://fetch.spec.whatwg.org/#request-body-header-name

Functions

append_a_request_origin_header ğŸ”’

https://fetch.spec.whatwg.org/#append-a-request-origin-header

append_the_fetch_metadata_headers ğŸ”’

https://w3c.github.io/webappsec-fetch-metadata/#abstract-opdef-append-the-fetch-metadata-headers-for-a-request

auth_from_cache ğŸ”’

block_for_cache_ready ğŸ”’

If the cache is not ready to construct a response, wait.

build_tls_security_info ğŸ”’

cors_check ğŸ”’

CORS check

cors_preflight_fetch ğŸ”’

CORS preflight fetch

cross_origin_resource_policy_check ğŸ”’

determine_requests_referrer

https://w3c.github.io/webappsec-referrer-policy/#determine-requests-referrer

has_credentials ğŸ”’

http_fetch

HTTP fetch

http_network_fetch ğŸ”’

HTTP network fetch

http_network_or_cache_fetch ğŸ”’

HTTP network or cache fetch

http_redirect_fetch

HTTP redirect fetch

is_no_store_cache ğŸ”’

is_redirect_status ğŸ”’

https://fetch.spec.whatwg.org/#redirect-status

is_schemelessy_same_site ğŸ”’

https://html.spec.whatwg.org/multipage/#schemelessly-same-site

location_url_for_response ğŸ”’

https://fetch.spec.whatwg.org/#concept-response-location-url

no_referrer_when_downgrade ğŸ”’

https://w3c.github.io/webappsec-referrer-policy/#referrer-policy-state-no-referrer-when-downgrade

obtain_response ğŸ”’

origin_when_cross_origin ğŸ”’

https://w3c.github.io/webappsec-referrer-policy/#referrer-policy-origin-when-cross-origin

prepare_devtools_request ğŸ”’

same_origin ğŸ”’

https://w3c.github.io/webappsec-referrer-policy/#referrer-policy-same-origin

send_early_httprequest_to_devtools

send_request_to_devtools

send_response_to_devtools

send_response_values_to_devtools

send_security_info_to_devtools

serialize_origin

Step 3 of https://fetch.spec.whatwg.org/#serializing-a-request-origin.

serialize_request_origin ğŸ”’

https://fetch.spec.whatwg.org/#serializing-a-request-origin

set_cookie_for_url ğŸ”’

set_cookies_from_headers ğŸ”’

set_default_accept ğŸ”’

Step 11 of https://fetch.spec.whatwg.org/#concept-fetch.

set_default_accept_encoding ğŸ”’

set_request_cookies ğŸ”’

set_requests_referrer_policy_on_redirect ğŸ”’

https://w3c.github.io/webappsec-referrer-policy/#set-requests-referrer-policy-on-redirect

set_the_sec_fetch_dest_header ğŸ”’

https://w3c.github.io/webappsec-fetch-metadata/#abstract-opdef-set-dest

set_the_sec_fetch_mode_header ğŸ”’

https://w3c.github.io/webappsec-fetch-metadata/#abstract-opdef-set-mode

set_the_sec_fetch_site_header ğŸ”’

https://w3c.github.io/webappsec-fetch-metadata/#abstract-opdef-set-site

set_the_sec_fetch_user_header ğŸ”’

https://w3c.github.io/webappsec-fetch-metadata/#abstract-opdef-set-user

strict_origin ğŸ”’

https://w3c.github.io/webappsec-referrer-policy/#referrer-policy-strict-origin

strict_origin_when_cross_origin ğŸ”’

https://w3c.github.io/webappsec-referrer-policy/#referrer-policy-strict-origin-when-cross-origin

strip_url_for_use_as_referrer ğŸ”’

https://w3c.github.io/webappsec-referrer-policy/#strip-url

wait_for_inflight_requests ğŸ”’

Wait for a cached response from channel. Happens when a fetch gets a cache hit, and the resource is pending completion from the network.