Module net::http_loader

Structs

HANDLE

HttpState

RedirectEndTimer

ResponseEndTimer

Enums

BodyChunk
Messages from the IPC route to the fetch worker, used to fill the body with bytes coming-in over IPC.

BodySink
The sink side of the body passed to hyper, used to enqueue chunks.

BodyStream
The stream side of the body passed to hyper.

HttpCacheEntryState
The various states an entry of the HttpCache can be in.

Functions

auth_from_cache

cors_check
CORS check

cors_preflight_fetch
CORS preflight fetch

determine_requests_referrer
https://w3c.github.io/webappsec-referrer-policy/#determine-requests-referrer

has_credentials

http_fetch
HTTP fetch

http_network_fetch
HTTP network fetch

http_network_or_cache_fetch
HTTP network or cache fetch

http_redirect_fetch
HTTP redirect fetch

is_no_store_cache

is_redirect_status
https://fetch.spec.whatwg.org/#redirect-status

is_schemelessy_same_site
https://html.spec.whatwg.org/multipage/#schemelessly-same-site

no_referrer_when_downgrade
https://w3c.github.io/webappsec-referrer-policy/#referrer-policy-state-no-referrer-when-downgrade

obtain_response

origin_when_cross_origin
https://w3c.github.io/webappsec-referrer-policy/#referrer-policy-origin-when-cross-origin

precise_time_ms

prepare_devtools_request

same_origin
https://w3c.github.io/webappsec-referrer-policy/#referrer-policy-same-origin

send_request_to_devtools

send_response_to_devtools

set_cookie_for_url

set_cookies_from_headers

set_default_accept

set_default_accept_encoding

set_default_accept_language

set_request_cookies

strict_origin
https://w3c.github.io/webappsec-referrer-policy/#referrer-policy-strict-origin

strict_origin_when_cross_origin
https://w3c.github.io/webappsec-referrer-policy/#referrer-policy-strict-origin-when-cross-origin

strip_url_for_use_as_referrer
https://w3c.github.io/webappsec-referrer-policy/#strip-url

try_immutable_origin_to_hyper_origin