net

Module image_cache

Structs

AllPendingLoads 🔒

CompletedLoad 🔒

Represents an image that has completed loading. Images that fail to load (due to network or decode failure) are still stored here, so that they aren’t fetched again.

DecoderMsg 🔒

Message that the decoder worker threads send to the image cache.

ImageCacheFactoryImpl

ImageCacheImpl

ImageCacheStore 🔒

Image cache implementation.

KeyCache 🔒

As getting new keys takes a round trip over the constellation, we keep a small cache of them. Additionally, this cache will store image resources that do not have a key yet because those are needed to complete the load.

LoadKeyGenerator 🔒

PendingLoad 🔒

Represents an image that is either being loaded by the resource thread, or decoded by a worker thread.

RasterizationTask 🔒

VectorImageData 🔒

Enums

CacheResult 🔒

Result of accessing a cache.

DecodedImage 🔒

ImageBytes 🔒

KeyCacheState 🔒

The state of the `WebRenderImageKey`` cache

LoadResult 🔒

PendingKey 🔒

Used for storing images that do not have a WebRenderImageKey yet.

Constants

FALLBACK_RIPPY 🔒

MAX_SVG_PIXMAP_DIMENSION 🔒

The current SVG stack relies on resvg to provide the natural dimensions of the SVG, which it automatically infers from the width/height/viewBox properties of the SVG. Since these can be arbitrarily large, this can cause us to allocate a pixmap with very large dimensions leading to the process being killed due to memory exhaustion. For example, the /css/css-transforms/perspective-svg-001.html test uses very large values for viewBox. Hence, we just clamp the maximum width/height of the pixmap allocated for rasterization.

Functions

decode_bytes_sync 🔒

parse_svg_document_in_memory 🔒

set_webrender_image_key 🔒

Type Aliases

ImageKey 🔒

https://html.spec.whatwg.org/multipage/#list-of-available-images

LoadKey 🔒