Module net::image_cache

Structs

AllPendingLoads

CompletedLoad

Represents an image that has completed loading. Images that fail to load (due to network or decode failure) are still stored here, so that they aren’t fetched again.

DecoderMsg

Message that the decoder worker threads send to the image cache.

ImageCacheImpl

ImageCacheStore

LoadKeyGenerator

PendingLoad

Represents an image that is either being loaded by the resource thread, or decoded by a worker thread.

Enums

CacheResult

Result of accessing a cache.

ImageBytes

LoadResult

Functions

decode_bytes_sync

TODO(gw): Remaining work on image cache: * Make use of the prefetch support in various parts of the code. * Profile time in GetImageIfAvailable - might be worth caching these results per paint / layout thread.

get_placeholder_image

set_webrender_image_key

Type Definitions

ImageKey

https://html.spec.whatwg.org/multipage/#list-of-available-images

LoadKey