Struct net::image_cache_thread::PendingLoad

struct PendingLoad {
    bytes: Vec<u8>,
    metadata: Option<ImageMetadata>,
    result: Option<Result<(), NetworkError>>,
    listeners: Vec<ImageListener>,
    url: Arc<Url>,
}

TODO(gw): Remaining work on image cache: * Make use of the prefetch support in various parts of the code. * Profile time in GetImageIfAvailable - might be worth caching these results per paint / layout thread.

MAYBE(Yoric): * For faster lookups, it might be useful to store the LoadKey in the DOM once we have performed a first load. Represents an image that is either being loaded by the resource thread, or decoded by a worker thread.

Fields
bytes: Vec<u8> metadata: Option<ImageMetadata> result: Option<Result<(), NetworkError>> listeners: Vec<ImageListener> url: Arc<Url>
Methods

impl PendingLoad

fn new(url: Arc<Url>) -> PendingLoad

fn add_listener(&mut self, listener: ImageListener)