net

Module protocols

Modules

blob 🔒

data 🔒

file 🔒

Structs

ProtocolRegistry

WebPageContentProtocolHandler 🔒

Enums

ProtocolRegisterError

Statics

FORBIDDEN_SCHEMES 🔒

Traits

ProtocolHandler

Functions

get_range_request_bounds

Get the range bounds if the Range header is present.

is_url_potentially_trustworthy

Test if the URL is potentially trustworthy or the custom protocol is registered as secure

partial_content

range_not_satisfiable_error

Type Aliases

FutureResponse 🔒