net::websocket_loader

Function run_ws_loop

async fn run_ws_loop(
    dom_receiver: UnboundedReceiver<DomMsg>,
    resource_event_sender: IpcSender<WebSocketNetworkEvent>,
    stream: WebSocketStream<ConnectStream>,
)

Listen for WS events from the DOM and the network until one side closes the connection or an error occurs. Since this is an async function that uses the select operation, it will run as a task on the WS tokio runtime.