Trait net::wire::pretty_print::PrettyPrint

pub trait PrettyPrint {
    // Required method
    fn pretty_print(
        buffer: &dyn AsRef<[u8]>,
        fmt: &mut Formatter<'_>,
        indent: &mut PrettyIndent
    ) -> Result<(), Error>;
}

Interface for printing listings.

Required Methods

fn pretty_print( buffer: &dyn AsRef<[u8]>, fmt: &mut Formatter<'_>, indent: &mut PrettyIndent ) -> Result<(), Error>

Write a concise, formatted representation of a packet contained in the provided buffer, and any nested packets it may contain.

pretty_print accepts a buffer and not a packet wrapper because the packet might be truncated, and so it might not be possible to create the packet wrapper.